+++ lib/vector-core/buffers/Cargo.toml
@@ -16,13 +16,13 @@ serde = { version = "1.0.126", default-features = false, features = ["derive"] }
 snafu = { version = "0.6.10", default-features = false, features = ["std"], optional = true }
 tokio = { version = "1.5.0", default-features = false, features = ["rt", "macros", "rt-multi-thread"] }
 tracing = { version = "0.1.26", default-features = false }
-quickcheck = { version = "1.0", default-features = false }
 
 [dev-dependencies]
-pretty_assertions = "0.7.2"
-tokio-test = "0.4.2"
 criterion = { version = "0.3", features = ["html_reports"] }
-tempdir = { version = "0.3", default-features = false }
+pretty_assertions = "0.7.2"
+quickcheck = "1.0"
+tempdir = "0.3"
+tokio-test = "0.4.2"
 
 [features]
 disk-buffer = ["db-key", "snafu", "leveldb"]
